requirements.txt
absl-py==0.9.0
argcomplete==1.10.0
astor==0.8.1
attrs==19.3.0
beautifulsoup4==4.8.0
#blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cymem==2.0.3
docx2txt==0.8
EbookLib==0.17.1
#en-core-web-sm==2.2.5
extract-msg==0.23.1
Flask==1.1.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
#grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
IMAPClient==2.1.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
langdetect==1.0.7
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
more-itertools==8.2.0
murmurhash==1.0.2
nltk==3.5
#numpy==1.18.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pdfminer.six==20181108
Pillow==7.1.1
plac==0.9.6
pluggy==0.13.1
preshed==3.0.2
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==2.0
spacy==2.2.4
spacy-langdetect==0.1.2
SpeechRecognition==3.8.1
srsly==1.0.2
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorflow-hub==0.8.0
tensorflow-text==1.15.0
termcolor==1.1.0
textract==1.6.3
thinc==7.4.0
tqdm==4.45.0
tzlocal==1.5.1
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==3.1.0
fuzzywuzzy==0.18.0
phonenumbers==8.12.12
Flask-Cors==3.0.9
ezodf==0.3.2